Clean-up hooks on plugin unload
