+++ core/metamod_plugins.h
@@ -39,6 +39,7 @@
 #include "IPluginManager.h"
 #include "ISmmPluginExt.h"
 #include "metamod_oslink.h"
+#include "metamod_khook.h"
 
 #include <string>
 #include <list>
@@ -95,6 +96,7 @@ public:
 		std::list<ProviderConCommand *> m_Cmds;
 		std::list<IMetamodListener *> m_Events;
 		METAMOD_FN_UNLOAD m_UnloadFn;
+		KHookImpl m_khook;
 	};
 public:
 	CPluginManager();
